+++ tests/compile-fail/normalized_state_collision.rs
@@ -1,0 +1,12 @@
+extern crate sml;
+
+use sml::sml;
+
+sml! {
+    _ {
+        *"foo-bar"_s + Start = Ready,
+        FooBar + Stop = Done,
+    }
+}
+
+fn main() {}